refactor: split `Interfaces.ts` into separate files
