+++ src/enums/ECurrentFlag.ts
@@ -1,0 +1,6 @@
+
+export enum ECurrentFlag {
+    None,
+    TestCase,
+    TestClass
+}
