+++ src/enums/ETestStatus.ts
@@ -1,0 +1,6 @@
+export enum ETestStatus {
+    Passed,
+    Failed,
+    Ignored,
+    Unknown
+}
