+++ src/enums/ETestType.ts
@@ -1,0 +1,5 @@
+export enum ETestType {
+    All,
+    Class,
+    Method
+}
