+++ src/gd/GdTestFileParser.ts
@@ -1,8 +1,8 @@
-import { group } from "console";
-import { ETestStatus, ITestCase, ITestClass } from "../Interfaces";
-import * as vscode from 'vscode';
+import { ITestClass } from "../interfaces/ITestClass";
+import { ITestCase } from "../interfaces/ITestCase";
+import { ETestStatus } from "../enums/ETestStatus";
 
-export const GdParseFile = (text: string) => {
+export const GdParseFile = (text: string): ITestClass | undefined => {
     const TestClassRe = /class_name(\s)+(?<className>\w+)(\s)+extends TestClass/m;
     const TestMethodRe = /func(\s)*(?<methodName>\w+)(\s)*\b\(\)/mg;
 
@@ -23,7 +23,7 @@ export const GdParseFile = (text: string) => {
     let match;
     while ((match = TestMethodRe.exec(text)) !== null) {
         if (!match.groups) {
-            return;
+            continue;
         }
         const testItem: ITestCase = {itemName: match.groups.methodName};
         TestFile.tests.push(testItem);
