+++ src/interfaces/ITestCase.ts
@@ -1,0 +1,7 @@
+import { ETestStatus } from "../enums/ETestStatus";
+
+export interface ITestCase {
+    itemName: string;
+    status?: ETestStatus;
+    args?: string | undefined;
+}
