+++ src/interfaces/ITestClass.ts
@@ -1,0 +1,6 @@
+import { ITestCase } from "./ITestCase";
+
+export interface ITestClass {
+    className: string;
+    tests: ITestCase[];
+}
